Transformed printing methods to instance methods.
As discussed in https://github.com/jmxtrans/jmxtrans/pull/163#discussion_r17832717.
