+++ src/com/googlecode/jmxtrans/example/Ehcache.java
@@ -6,7 +6,7 @@ import com.googlecode.jmxtrans.model.Query;
 import com.googlecode.jmxtrans.model.Server;
 import com.googlecode.jmxtrans.model.output.GraphiteWriter;
 import com.googlecode.jmxtrans.util.BaseOutputWriter;
-import com.googlecode.jmxtrans.util.JsonUtils;
+import com.googlecode.jmxtrans.util.JsonPrinter;
 
 /**
  * This example shows how to query ehcache for its statistics information.
@@ -16,6 +16,7 @@ import com.googlecode.jmxtrans.util.JsonUtils;
 public class Ehcache {
 
 	private static final String GW_HOST = "192.168.192.133";
+	private static final JsonPrinter printer = new JsonPrinter(System.out);
 
 	/** */
 	public static void main(String[] args) throws Exception {
@@ -45,7 +46,7 @@ public class Ehcache {
 		server.addQuery(q);
 
 		JmxProcess process = new JmxProcess(server);
-		JsonUtils.prettyPrintJson(process, System.out);
+		printer.prettyPrint(process);
 
 		JmxTransformer transformer = new JmxTransformer();
 		transformer.executeStandalone(process);
