+++ src/com/googlecode/jmxtrans/example/InterestingInfo.java
@@ -6,7 +6,7 @@ import com.googlecode.jmxtrans.model.Query;
 import com.googlecode.jmxtrans.model.Server;
 import com.googlecode.jmxtrans.model.output.GraphiteWriter;
 import com.googlecode.jmxtrans.model.output.StdOutWriter;
-import com.googlecode.jmxtrans.util.JsonUtils;
+import com.googlecode.jmxtrans.util.JsonPrinter;
 
 /**
  * This class produces the json that is in example.json.
@@ -14,6 +14,8 @@ import com.googlecode.jmxtrans.util.JsonUtils;
  * @author jon
  */
 public class InterestingInfo {
+
+	private static final JsonPrinter printer = new JsonPrinter(System.out);
 
 	/** */
 	public static void main(String[] args) throws Exception {
@@ -57,7 +59,7 @@ public class InterestingInfo {
 		server.addQuery(q4);
 
 		JmxProcess process = new JmxProcess(server);
-		JsonUtils.prettyPrintJson(process, System.out);
+		printer.prettyPrint(process);
 		JmxTransformer transformer = new JmxTransformer();
 		transformer.executeStandalone(process);
 	}
