+++ src/com/googlecode/jmxtrans/example/Tester.java
@@ -5,7 +5,7 @@ import com.googlecode.jmxtrans.model.JmxProcess;
 import com.googlecode.jmxtrans.model.Query;
 import com.googlecode.jmxtrans.model.Server;
 import com.googlecode.jmxtrans.model.output.StdOutWriter;
-import com.googlecode.jmxtrans.util.JsonUtils;
+import com.googlecode.jmxtrans.util.JsonPrinter;
 
 /**
  * This class produces the json that is in example.json.
@@ -13,6 +13,8 @@ import com.googlecode.jmxtrans.util.JsonUtils;
  * @author jon
  */
 public class Tester {
+
+	private static final JsonPrinter printer = new JsonPrinter(System.out);
 
 	/** */
 	public static void main(String[] args) throws Exception {
@@ -38,7 +40,7 @@ public class Tester {
 		server.addQuery(q3);
 
 		JmxProcess process = new JmxProcess(server);
-		JsonUtils.prettyPrintJson(process, System.out);
+		printer.prettyPrint(process);
 		JmxTransformer transformer = new JmxTransformer();
 		transformer.executeStandalone(process);
 	}
